requirements.txt
attrs==21.4.0
certifi==2022.5.18.1
cfgv==3.3.1
charset-normalizer==2.0.12
distlib==0.3.6
ecdsa==0.18.0
filelock==3.9.0
identify==2.5.15
idna==3.3
iniconfig==1.1.1
nodeenv==1.7.0
packaging==21.3
parsedatetime==2.6
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.0
py==1.11.0
pyasn1==0.4.8
pyparsing==3.0.9
pytest==7.1.2
pytest-check==1.3.0
pytest-testconfig==0.2.0
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak==2.7.0
PyYAML==6.0
requests==2.28.0
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
virtualenv==20.17.1
wait-for==1.2.0
